+++ Changelog.md
@@ -11,3 +11,18 @@ Contains notable changes there were added, fixed, or removed in each release.
 * Build Board source as a library
 * Template to build arbitrary number of targets
 * Auto-generating documentation template
+
+## 0.2.0
+
+### Changes
+
+* ADS8689IPWR driver created and implemented
+* RedundantADC driver created and implemented
+* HIB driver created and implemented
+* Updated README.md
+* Created targets for HIB, RedundantADC testing, and ADS88689IPWR testing
+* Improved dependency order
+* Verified ADC results
+* Added error checking for ADCs
+* Added logger statements for easier debugging
+* Documentation added for each class as well as each target